requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
aiohttp-socks==0.6.0
async-timeout==3.0.1
attr==0.3.1
attrs==21.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
cchardet==2.1.7
certifi==2020.6.20
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
click==8.0.1
dataclasses==0.6
elasticsearch==7.13.1
fake-useragent==0.1.11
geographiclib==1.50
geopy==2.1.0
googletransx==2.4.2
idna==3.2
joblib==1.0.1
multidict==5.1.0
nest-asyncio==1.5.1
nltk==3.6.2
numpy==1.21.1
pandas==1.3.1
pycares==4.0.0
pycparser==2.20
PySocks==1.7.1
python-dateutil==2.8.2
python-socks==1.2.4
pytz==2021.1
regex==2021.7.6
requests==2.26.0
schedule==1.1.0
six==1.16.0
soupsieve==2.2.1
textblob==0.15.3
tqdm==4.61.2
tweet-preprocessor==0.6.0
twint @ git+https://github.com/FedericoMz/twint.git@8d2384d485df85a84c77e7711744e5bff409d1d0
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3
